+++ src/test/java/org/wickedsource/docxstamper/ConditionalDisplayOfParagraphsTest.java
@@ -10,6 +10,7 @@ import org.docx4j.wml.Tr;
 import org.junit.jupiter.api.Test;
 import org.wickedsource.docxstamper.util.DocumentUtil;
 import org.wickedsource.docxstamper.util.ParagraphWrapper;
+import pro.verron.docxstamper.utils.TestDocxStamper;
 
 import java.io.IOException;
 import java.util.List;
