+++ src/test/java/org/wickedsource/docxstamper/ConditionalDisplayOfTableRowsTest.java
@@ -8,6 +8,7 @@ import org.docx4j.wml.Tc;
 import org.docx4j.wml.Tr;
 import org.junit.jupiter.api.Test;
 import org.wickedsource.docxstamper.util.DocumentUtil;
+import pro.verron.docxstamper.utils.TestDocxStamper;
 
 import java.io.IOException;
 import java.util.List;
