+++ src/test/java/org/wickedsource/docxstamper/ConditionalDisplayOfTablesBug32Test.java
@@ -9,6 +9,7 @@ import org.docx4j.wml.Tc;
 import org.docx4j.wml.Tr;
 import org.junit.jupiter.api.Test;
 import org.wickedsource.docxstamper.util.ParagraphWrapper;
+import pro.verron.docxstamper.utils.TestDocxStamper;
 
 import java.io.IOException;
