+++ src/test/java/org/wickedsource/docxstamper/CustomTypeResolverTest.java
@@ -5,6 +5,7 @@ import org.docx4j.wml.P;
 import org.junit.jupiter.api.Test;
 import org.wickedsource.docxstamper.replace.typeresolver.AbstractToTextResolver;
 import org.wickedsource.docxstamper.util.ParagraphWrapper;
+import pro.verron.docxstamper.utils.TestDocxStamper;
 
 import java.io.IOException;
