+++ src/test/java/org/wickedsource/docxstamper/DateReplacementTest.java
@@ -3,8 +3,8 @@ package org.wickedsource.docxstamper;
 import org.docx4j.openpackaging.exceptions.Docx4JException;
 import org.docx4j.wml.P;
 import org.junit.jupiter.api.Test;
-import org.wickedsource.docxstamper.context.DateContext;
 import org.wickedsource.docxstamper.util.ParagraphWrapper;
+import pro.verron.docxstamper.utils.TestDocxStamper;
 
 import java.io.IOException;
 import java.text.SimpleDateFormat;
@@ -13,6 +13,19 @@ import java.util.Date;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class DateReplacementTest {
+	public static class DateContext {
+
+		private Date date;
+
+		public Date getDate() {
+			return date;
+		}
+
+		public void setDate(Date date) {
+			this.date = date;
+		}
+	}
+
 	@Test
 	public void test() throws Docx4JException, IOException {
 		var now = new Date();
