+++ src/test/java/org/wickedsource/docxstamper/ExpressionReplacementInHeaderAndFooterTest.java
@@ -10,6 +10,7 @@ import org.docx4j.relationships.Relationship;
 import org.docx4j.wml.P;
 import org.junit.jupiter.api.Test;
 import org.wickedsource.docxstamper.util.ParagraphWrapper;
+import pro.verron.docxstamper.utils.TestDocxStamper;
 
 import java.io.IOException;
