+++ src/test/java/org/wickedsource/docxstamper/ExpressionReplacementInTextBoxesTest.java
@@ -3,6 +3,7 @@ package org.wickedsource.docxstamper;
 import org.docx4j.dml.wordprocessingDrawing.Anchor;
 import org.docx4j.openpackaging.exceptions.Docx4JException;
 import org.junit.jupiter.api.Test;
+import pro.verron.docxstamper.utils.TestDocxStamper;
 
 import java.io.IOException;
 import java.util.List;
