+++ src/test/java/org/wickedsource/docxstamper/ExpressionReplacementWithFormattingTest.java
@@ -2,6 +2,7 @@ package org.wickedsource.docxstamper;
 
 import org.docx4j.openpackaging.exceptions.Docx4JException;
 import org.junit.jupiter.api.Test;
+import pro.verron.docxstamper.utils.TestDocxStamper;
 
 import java.io.IOException;
 import java.util.List;
