+++ src/test/java/org/wickedsource/docxstamper/RepeatDocPartAndCommentProcessorsIsolationTest.java
@@ -2,6 +2,7 @@ package org.wickedsource.docxstamper;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.context.expression.MapAccessor;
+import pro.verron.docxstamper.utils.TestDocxStamper;
 
 import java.util.ArrayList;
 import java.util.HashMap;
