+++ src/test/java/org/wickedsource/docxstamper/RepeatDocPartWithImageTest.java
@@ -3,6 +3,7 @@ package org.wickedsource.docxstamper;
 import org.junit.jupiter.api.Test;
 import org.springframework.context.expression.MapAccessor;
 import org.wickedsource.docxstamper.replace.typeresolver.image.Image;
+import pro.verron.docxstamper.utils.TestDocxStamper;
 
 import java.io.IOException;
 import java.util.List;
