+++ src/test/java/org/wickedsource/docxstamper/RepeatParagraphTest.java
@@ -4,21 +4,20 @@ import org.docx4j.openpackaging.exceptions.Docx4JException;
 import org.docx4j.openpackaging.packages.WordprocessingMLPackage;
 import org.docx4j.wml.P;
 import org.junit.jupiter.api.Test;
-import org.wickedsource.docxstamper.context.Character;
-import org.wickedsource.docxstamper.context.CharactersContext;
 import org.wickedsource.docxstamper.util.ParagraphWrapper;
 import org.wickedsource.docxstamper.util.walk.BaseCoordinatesWalker;
+import pro.verron.docxstamper.utils.TestDocxStamper;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class RepeatParagraphTest {
-
 	@Test
 	public void test() throws Docx4JException, IOException {
-		var context = new CharactersContext();
+		var context = new Characters();
 		context.getCharacters().add(new Character("Homer Simpson", "Dan Castellaneta"));
 		context.getCharacters().add(new Character("Marge Simpson", "Julie Kavner"));
 		context.getCharacters().add(new Character("Bart Simpson", "Nancy Cartwright"));
@@ -26,7 +25,7 @@ public class RepeatParagraphTest {
 		context.getCharacters().add(new Character("Disco Stu", "Hank Azaria"));
 		context.getCharacters().add(new Character("Krusty the Clown", "Dan Castellaneta"));
 		var template = getClass().getResourceAsStream("RepeatParagraphTest.docx");
-		var stamper = new TestDocxStamper<CharactersContext>();
+		var stamper = new TestDocxStamper<Characters>();
 		WordprocessingMLPackage document = stamper.stampAndLoad(template, context);
 
 		var titleCoords = new ArrayList<P>();
@@ -59,6 +58,35 @@ public class RepeatParagraphTest {
 		assertEquals("Hank Azaria", new ParagraphWrapper(quotationCoords.get(4)).getText());
 		assertEquals("Krusty the Clown", new ParagraphWrapper(titleCoords.get(5)).getText());
 		assertEquals("Dan Castellaneta", new ParagraphWrapper(quotationCoords.get(5)).getText());
+	}
+
+	public static class Character {
+
+		private final String name;
+
+		private final String actor;
+
+		public Character(String name, String actor) {
+			this.name = name;
+			this.actor = actor;
+		}
+
+		public String getName() {
+			return name;
+		}
+
+		public String getActor() {
+			return actor;
+		}
+	}
+
+	public class Characters {
+
+		private final List<Character> characters = new ArrayList<>();
+
+		public List<Character> getCharacters() {
+			return characters;
+		}
 	}
