--- src/test/java/org/wickedsource/docxstamper/context/AClass.java
@@ -1,6 +1,0 @@
-package org.wickedsource.docxstamper.context;
-
-import java.util.List;
-
-public record AClass(int number, List<Student> students) {
-}
