--- src/test/java/org/wickedsource/docxstamper/context/CharactersContext.java
@@ -1,13 +1,0 @@
-package org.wickedsource.docxstamper.context;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class CharactersContext {
-
-    private final List<Character> characters = new ArrayList<>();
-
-    public List<Character> getCharacters() {
-        return characters;
-    }
-}
