--- src/test/java/org/wickedsource/docxstamper/context/DateContext.java
@@ -1,16 +1,0 @@
-package org.wickedsource.docxstamper.context;
-
-import java.util.Date;
-
-public class DateContext {
-
-    private Date date;
-
-    public Date getDate() {
-        return date;
-    }
-
-    public void setDate(Date date) {
-        this.date = date;
-    }
-}
