--- src/test/java/org/wickedsource/docxstamper/context/ImageContext.java
@@ -1,16 +1,0 @@
-package org.wickedsource.docxstamper.context;
-
-import org.wickedsource.docxstamper.replace.typeresolver.image.Image;
-
-public class ImageContext {
-
-    private Image monalisa;
-
-    public Image getMonalisa() {
-        return monalisa;
-    }
-
-    public void setMonalisa(Image monalisa) {
-        this.monalisa = monalisa;
-    }
-}
