--- src/test/java/org/wickedsource/docxstamper/context/SchoolContext.java
@@ -1,25 +1,0 @@
-package org.wickedsource.docxstamper.context;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class SchoolContext {
-    public SchoolContext() {
-    }
-
-    String schoolName;
-    List<Grade> grades = new ArrayList<>();
-
-    public SchoolContext(String schoolName) {
-        this.schoolName = schoolName;
-    }
-
-    public String getSchoolName() {
-        return schoolName;
-    }
-
-    public List<Grade> getGrades() {
-        return grades;
-    }
-
-}
