+++ src/test/java/org/wickedsource/docxstamper/util/RunUtilTest.java
@@ -5,7 +5,7 @@ import org.docx4j.openpackaging.packages.WordprocessingMLPackage;
 import org.docx4j.wml.P;
 import org.docx4j.wml.R;
 import org.junit.jupiter.api.Test;
-import org.wickedsource.docxstamper.IOStreams;
+import pro.verron.docxstamper.utils.IOStreams;
 
 import java.io.IOException;
