+++ src/test/java/pro/verron/docxstamper/utils/RunCollector.java
@@ -1,4 +1,4 @@
-package org.wickedsource.docxstamper.util;
+package pro.verron.docxstamper.utils;
 
 import org.docx4j.utils.TraversalUtilVisitor;
 import org.docx4j.wml.R;
