+++ src/test/java/pro/verron/docxstamper/utils/Stringifier.java
@@ -1,4 +1,4 @@
-package pro.verron.docxstamper;
+package pro.verron.docxstamper.utils;
 
 import jakarta.xml.bind.JAXBElement;
 import lombok.SneakyThrows;
@@ -13,7 +13,6 @@ import org.docx4j.dml.wordprocessingDrawing.Inline;
 import org.docx4j.openpackaging.packages.WordprocessingMLPackage;
 import org.docx4j.openpackaging.parts.WordprocessingML.BinaryPartAbstractImage;
 import org.docx4j.wml.*;
-import org.wickedsource.docxstamper.util.RunCollector;
 
 import java.security.MessageDigest;
 import java.text.CharacterIterator;
